requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pytest
google-generativeai
python-dotenv
gunicorn
selenium
webdriver-manager
Flask-JWT-Extended